rename functions to past tense
