+++ bigbluebutton-html5/imports/api/2.0/users/server/handlers/userKicked.js
@@ -1,7 +1,7 @@
-import kickUser from '../modifiers/kickUser';
+import userKicked from '../modifiers/userKicked';
 
 export default function handleKickUser({ header }) {
   const { meetingId, userId } = header;
 
-  return kickUser(meetingId, userId);
+  return userKicked(meetingId, userId);
 }
