+++ bigbluebutton-html5/imports/api/2.0/users/server/modifiers/userKicked.js
